requirements.txt
pandas
snowflake-connector-python
python-dotenv
pyspark
kagglehub
